+++ src/Scripts/FunctionMappings/ThanatosLogic.lua
@@ -248,6 +248,9 @@ function mod.HandleThanatosEncounterReward(thanatos, args)
 		end
 	end
 
+	-- Need to move his kills into CurrentRun as otherwise they are not kept in the save
+	game.CurrentRun.ModsNikkelMHadesBiomesTotalThanatosKills = (game.CurrentRun.ModsNikkelMHadesBiomesTotalThanatosKills or 0) + (encounter.ThanatosKills or 0)
+
 	game.NPCRewardDropPreProcessArgs(thanatos.KillChallengeArgs)
 	game.NPCRewardDrop(thanatos, thanatos.KillChallengeArgs)